requirements.txt
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
fastapi==0.89.1
h11==0.14.0
idna==3.4
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.4
Pygments==2.14.0
python-dotenv==0.21.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.22.0
traitlets==5.8.1
typer==0.7.0
typing_extensions==4.4.0
uvicorn==0.20.0
wcwidth==0.2.5
